+++ packages/const/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@lobechat/const",
+  "version": "1.0.0",
+  "private": true,
+  "exports": {
+    "./*": {
+      "types": [
+        "./src/*.ts",
+        "./src/*/index.ts"
+      ],
+      "default": [
+        "./src/*.ts",
+        "./src/*/index.ts"
+      ]
+    }
+  }
+}
