+++ packages/const/src/auth.ts
