+++ packages/const/src/guide.ts
