+++ packages/const/src/image.ts
