+++ packages/const/src/settings/hotkey.ts
