+++ packages/const/src/version.ts
