+++ tsconfig.json
@@ -20,6 +20,7 @@
     "paths": {
       "@/libs/model-runtime": ["./packages/model-runtime/src/index.ts"],
       "@/libs/model-runtime/*": ["./packages/model-runtime/src/*"],
+      "@/const/*": ["./packages/const/src/*"],
       "@/types/*": ["./packages/types/src/*", "./src/types/*"],
       "@/*": ["./src/*"],
       "~test-utils": ["./tests/utils.tsx"]
